Unit testing 96%

+++ src/main/java/com/revature/project2/services/ItemService.java
@@ -1,5 +1,6 @@
 package com.revature.project2.services;
 
+import com.revature.project2.exceptions.InvalidRequestException;
 import com.revature.project2.exceptions.ResourceNotFoundException;
 import com.revature.project2.models.Genre;
 import com.revature.project2.models.Item;
@@ -23,43 +24,92 @@ public class ItemService {
         this.itemRepo = itemRepository;
     }
 
-
-    @Transactional(readOnly=true)
-    public List<Item> findAll() {
-        return itemRepo.findAll();
-    }
-
+    /**
+     * CREATE operation
+     * @param newItem
+     * @return
+     */
     @Transactional
     public Item save(Item newItem) {
         return itemRepo.save(newItem).orElseThrow(ResourceNotFoundException::new);
     }
 
+    /**
+     * READ operation
+     * @return
+     */
+    @Transactional(readOnly=true)
+    public List<Item> findAll() {
+        return itemRepo.findAll();
+    }
+
+    /**
+     * READ operation
+     * @param id
+     * @return
+     */
     @Transactional
     public Item findById(Integer id) {
+
+        if (id <= 0) {
+            throw new InvalidRequestException("ID cannot be negative or equal to zero");
+        }
+
         return itemRepo.findById(id).orElseThrow(ResourceNotFoundException::new);
     }
 
-    @Transactional
-    public boolean update(Item updateItem) {
-        return itemRepo.update(updateItem);
-    }
-
-    @Transactional
-    public boolean deleteById(Integer id) {
-        return itemRepo.deleteById(id);
-    }
-
+    /**
+     * READ operation
+     * @param genre
+     * @return
+     */
     @Transactional
     public List<Item> findUsersByGenre(String genre) {
         Genre genreEnum = Genre.valueOf(genre.toUpperCase());
         return  itemRepo.findUsersByGenre(genreEnum);
     }
 
+    /**
+     * READ operation
+     * @param name
+     * @return
+     */
     @Transactional
     public Item findUserByName(String name){
         return itemRepo.findUserByName(name).orElseThrow(ResourceNotFoundException::new);
     }
 
+    /**
+     * UPDATE operation
+     * @param updateItem
+     * @return
+     */
+    @Transactional
+    public boolean update(Item updateItem) {
+        return itemRepo.update(updateItem);
+    }
+
+    /**
+     * DELETE operation
+     * @param id
+     * @return
+     */
+    @Transactional
+    public boolean deleteById(Integer id) {
+
+        if (id <= 0) {
+            throw new InvalidRequestException("ID cannot be negative or equal to zero");
+        }
+
+        return itemRepo.deleteById(id);
+    }
+
+
+    /**
+     * Convenience method
+     * @param name
+     * @return
+     */
     @Transactional
     public boolean isNameValid(String name){
         return itemRepo.isNameValid(name);
